requirements.txt
PyPDF2 
Pillow 
python-docx 
openpyxl 
python-pptx
